[STYLE] rename major file
