+++ functions/api/routes/major/index.js
@@ -37,7 +37,7 @@ const router = express.Router();
  *       '500':
  *         description: 서버 내부 에러
  */
-router.get("/university/:universityId", require("./majorListGET"));
+router.get("/university/:universityId", require("./majorUniversityListGET"));
 
 /**
  * @swagger
