+++ functions/api/routes/major/majorUniversityListGET.js
